# src/management-service/docs/swagger.yaml
definitions:
...
  request.CreateCategoryRequest:
    properties:
      categoryName:
        type: string
      description:
        type: string
    type: object
...
  response.Standard:
    properties:
      data: {}
      message:
        type: string
      statusCode:
        type: integer
    type: object
info:
  contact: {}
paths:
  /create-category:
    post:
      consumes:
      - application/json
      description: Create a new category with a name and description
      parameters:
      - description: Category creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Standard'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Standard'
        "409":
          description: Conflict (Duplicate Category)
          schema:
            $ref: '#/definitions/response.Standard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Standard'
      summary: Create a new category
      tags:
      - categories
...
  /delete-category/{id}:
    delete:
      description: Soft delete a category by its UUID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Standard'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/response.Standard'
        "404":
          description: Category Not Found
          schema:
            $ref: '#/definitions/response.Standard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Standard'
      summary: Delete a category by ID
      tags:
      - categories
...
  /get-all-categories:
    get:
      description: Get a list of all categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Standard'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Standard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Standard'
      summary: Get all categories
      tags:
      - categories
...
  /get-category-id/{id}:
    get:
      description: Get a single category by its UUID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Standard'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/response.Standard'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Standard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Standard'
      summary: Get category by ID
      tags:
      - categories
...
  /restore-category/{id}:
    post:
      description: Restore a category that was previously soft deleted by its UUID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Standard'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/response.Standard'
        "404":
          description: Category Not Found
          schema:
            $ref: '#/definitions/response.Standard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Standard'
      summary: Restore a soft deleted category by ID
      tags:
      - categories
...
  /update-category/{id}:
    put:
      consumes:
      - application/json
      description: Update a category's name and description by its UUID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Standard'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/response.Standard'
        "404":
          description: Category Not Found
          schema:
            $ref: '#/definitions/response.Standard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Standard'
      summary: Update a category by ID
      tags:
      - categories
...
swagger: "2.0"